fix: support nested blog slugs
